Separate utils and run scripts
