+++ local_group_support/post_signups_to_mattermost.py
@@ -1,41 +1,11 @@
 import datetime
-import os
 
-import requests
-from dotenv import load_dotenv
-from local_group_support.members import get_member_stats
-
-
-BASE_URL = 'https://organise.earth/'
-LOCAL_GROUP_INTEGRATORS_CHANNEL = 'nqs4h6iyrpr3mx4jjy9xqk8i3o'
-
-
-def get_mattermost_session_token():
-    load_dotenv()
-    key = os.getenv("MATTERMOST_TOKEN")
-
-    if not key:
-        raise OSError('MATTERMOST_TOKEN not found in .env')
-
-    return key
-
-
-def post_to_channel(channel_id, message):
-    data = {"channel_id": channel_id, "message": message}
-    headers = {'Authorization': get_mattermost_session_token()}
-    response = requests.post(BASE_URL + 'api/v4/posts', headers=headers, json=data)
-    return response
-
-
-def get_mattermost_user(email):
-    url = BASE_URL + f'/api/v4/users/email/{email}'
-    headers = {'Authorization': get_mattermost_session_token()}
-    response = requests.post(url, headers=headers)
-    if response.status_code == 200:
-        return response.json()
-
+from local_group_support.utils.members import get_member_stats
+from local_group_support.utils.mattermost import post_to_channel
 
 if __name__ == "__main__":
+    LOCAL_GROUP_INTEGRATORS_CHANNEL = 'nqs4h6iyrpr3mx4jjy9xqk8i3o'
+
     start_date = datetime.date.today() - datetime.timedelta(days=30)
     df = get_member_stats(start_date)
     df_grouped = df.groupby('local_group').size()
