+++ local_group_support/utils/members.py
@@ -3,8 +3,7 @@ import datetime
 import pandas as pd
 
 from local_group_support.config.config import get_config
-from local_group_support.forms import get_forms
-from local_group_support.util import query, query_all
+from local_group_support.utils.action_network import get_forms, query, query_all
 
 FORMATION_DATE = datetime.date(2018, 4, 1)
